refactor: rename ResetPassword endpoint to CreatePasswordResetToken
