+++ src/Human.WebServer.Api.V1/Auth/CreatePasswordResetToken/Response.cs
@@ -1,0 +1,15 @@
+﻿using Human.Core.Features.Auth.CreatePasswordResetToken;
+using Riok.Mapperly.Abstractions;
+
+namespace Human.WebServer.Api.V1.Auth.CreatePasswordResetToken;
+
+internal sealed class CreatePasswordResetTokenResponse
+{
+    public required string Token { get; set; }
+}
+
+[Mapper]
+internal static partial class ResponseMapper
+{
+    public static partial CreatePasswordResetTokenResponse ToResponse(this CreatePasswordResetTokenResult result);
+}
